// src/app/models/datosQr.model.ts
export interface datosQR {
    nombrePasajero: string;
    dniPasajero: string;
    numeroVuelo: string;
    numeroAsiento: string;
    clase: string;
    fechaVuelo: Date;
    aeropuertoSalida: string;
    ciudadSalida: string;
    horaSalida: Date;
    aeropuertoLlegada: string;
    ciudadLlegada: string;
    horaLlegada: Date;
    horaEmbarque: Date;
    puertaEmbarque: string;
}